+++ Src/Public/Footer.ps1
@@ -51,7 +51,7 @@ function Footer
             {
                 if (('Id' -in $result.PSObject.Properties.Name) -and
                     ('Type' -in $result.PSObject.Properties.Name) -and
-                    ($result.Type -in 'PScribo.Paragraph','PScribo.Table'))
+                    ($result.Type -in 'PScribo.Paragraph','PScribo.Table','PScribo.LineBreak'))
                 {
                     [ref] $null = $pscriboFooter.Sections.Add($result)
                 }